cleaned up dumb mistakes
