+++ include/PluckyInterface.hpp
@@ -16,36 +16,10 @@ public:
   virtual bool availableForWrite(size_t len=0) = 0;
   virtual bool writeAll(const uint8_t *buf, size_t size) = 0;
 
-protected:
-  char _interfaceName[32];
-
-
-  // helper functions 
-  void trimBuffer(uint8_t *buf, uint16_t &len) {
-    if (buf[len-1] == '\n') {
-      if (buf[len-2] == '\r') {
-        // convert CRLF to CR just to make everyone's lives easier
-        // but also log a complaint about it
-        buf[len-2] = '\n';
-        len = len-1;
-        Logger.warning.printf("Stripped CRLF from interface %s", _interfaceName);
-      }
-    }
-    if (len < READ_BUFFER_SIZE) {
-      buf[len] = 0; // force null termination for convenience
-    }
-  }
-
-  void debugHandler(uint8_t *buf, uint16_t &len) {
-    if (strncmp((char *)buf, "HEAP", 9)) {
-      Logger.debug.printf("Free Heap: %d", esp_get_free_heap_size());
-      len=0;
-    }
-  }
-
-
 };
 
-
+// helper functions 
+void trimBuffer(uint8_t *buf, uint16_t &len, char *interfaceName);
+void debugHandler(uint8_t *buf, uint16_t &len);
 
 #endif // _PLUCKY_INTERFACE_HPP_
