+++ src/PluckyInterfaceSerial.cpp
@@ -86,7 +86,7 @@ bool PluckyInterfaceSerial::readAll() {
             uint16_t sendLen = _readBufIndex;
             _readBufIndex = 0;
 
-            trimBuffer(_readBuf, sendLen);
+            trimBuffer(_readBuf, sendLen, _interfaceName);
             debugHandler(_readBuf, sendLen);
 
             if (_uart_nr == SERIAL_DE_UART_NUM) {
